@@ -2,6 +2,7 @@
 
 namespace Themsaid\CashierTool;
 
+use Illuminate\Http\Request;
 use Laravel\Nova\Nova;
 use Laravel\Nova\Tool;
 
@@ -14,7 +15,43 @@ class CashierTool extends Tool
      */
     public function boot()
     {
-        Nova::script('nova-cashier-tool', __DIR__.'/../dist/js/tool.js');
-        Nova::style('nova-cashier-tool', __DIR__.'/../dist/css/tool.css');
+        parent::boot();
+
+        Nova::script('nova-cashier-tool', __DIR__ . '/../dist/js/tool.js');
+        Nova::style('nova-cashier-tool', __DIR__ . '/../dist/css/tool.css');
+    }
+
+    /**
+     * Build the view that renders the navigation links for the tool.
+     *
+     * @return \Illuminate\View\View|string
+     */
+    public function renderNavigation()
+    {
+        return view('nova-cashier-tool::navigation');
+    }
+
+    /**
+     * Get the displayable name of the tool.
+     *
+     * @return string
+     */
+    public function name()
+    {
+        return __('Cashier Manager');
+    }
+
+    /**
+     * Get the tools that should be listed in the Nova sidebar.
+     *
+     * @return array
+     */
+    public function menu(Request $request)
+    {
+        return [
+            (new CashierResourceTool)->canSee(function ($request) {
+                return $request->user()->can('manage cashier');
+            }),
+        ];
     }
 }
